requirements.txt
gymnasium==0.29.1
torch>=2.1.0
numpy>=1.24,<2.0
PyYAML>=6.0
pygame>=2.5.2
matplotlib>=3.8.0
tqdm
